Use uint32_t explicitly for the total FW size

+++ components/ocs_system/crc32_updater.h
@@ -22,7 +22,7 @@ public:
     Crc32Updater(ICrc32Calculator& crc32_calculator, IUpdater& updater);
 
     //! Begin firmware update.
-    status::StatusCode begin(size_t total_size, uint32_t crc32) override;
+    status::StatusCode begin(uint32_t total_size, uint32_t crc32) override;
 
     //! Write firmware data.
     status::StatusCode write(const uint8_t* buf, size_t len) override;
