+++ components/ocs_system/target_esp32/partition_updater.h
@@ -26,7 +26,7 @@ public:
                      const char* partname);
 
     //! Begin firmware update.
-    status::StatusCode begin(size_t total_size, uint32_t crc32) override;
+    status::StatusCode begin(uint32_t total_size, uint32_t crc32) override;
 
     //! Write firmware data to the updating partition.
     status::StatusCode write(const uint8_t* buf, size_t len) override;
